feat(revise): handle operation interrupt
