+++ src/lib.rs
@@ -1,4 +1,3 @@
 pub mod commit;
 pub mod config;
 pub mod error;
-pub mod inquire;
